+++ changelog.md
@@ -7,6 +7,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [untagged]
 
 - /api/timesheets/ require header Authorization with an apikey
-- Added /api/timesheets/
+- Added /api/timesheets/{account}/yyyymm.timesheet
 
 ## [0.1.0] 2020-06-02